requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

openai==1.55.3
playwright==1.48.0
selenium==4.25.0
requests==2.31.0
beautifulsoup4==4.12.3

pandas==2.0.3
numpy==1.26.3
pillow==9.5.0

PyPDF2==3.0.1
pdfplumber==0.11.0
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.4.0

aiohttp==3.9.1
lxml==5.1.0
openpyxl==3.1.2
pytesseract==0.3.10
opencv-python-headless==4.9.0.80

gunicorn==21.2.0
